# Conflicts:
#	.vs/Massacre_Project/v17/.suo
#	.vs/Massacre_Project/v17/DocumentLayout.backup.json
#	.vs/Massacre_Project/v17/DocumentLayout.json
#	.vs/Massacre_Project/v17/fileList.bin
#	Assembly-CSharp.csproj
#	UserSettings/Layouts/default-2022.dwlt

+++ Assets/01.BSJ/02.Scripts/Monster/Data/IMonsterCombat.cs
@@ -4,8 +4,8 @@ public interface IMonsterCombat
     MonsterAttack MonsterAttack { get; }
     MonsterTargetDistance MonsterTargetDistance { get; }
 
-    int MoveSpeed { get; }
-    int TurnSpeed { get; }
+    float MoveSpeed { get; }
+    float TurnSpeed { get; }
 
     bool IsDead { get; set; }    // 생사 여부
 }
